requirements.txt
asgiref==3.7.2
Django==4.2.3
djangorestframework==3.15.2
pillow==11.0.0
psycopg2==2.9.10
sqlparse==0.4.4
tzdata==2023.3
